+++ app/src-tauri/src/main.rs
@@ -21,7 +21,7 @@ use nusamai::{
     source::{citygml::CityGmlSourceProvider, DataSourceProvider},
     transformer::{
         self, MappingRules, MultiThreadTransformer, NusamaiTransformBuilder, TransformBuilder,
-        TransformerRegistry,
+        TransformerSettings,
     },
 };
 use nusamai_plateau::models::TopLevelCityObject;
@@ -148,7 +148,7 @@ fn run_conversion(
     filetype: String,
     epsg: u16,
     rules_path: String,
-    transformer_registry: TransformerRegistry,
+    transformer_registry: TransformerSettings,
     sink_parameters: Parameters,
     tasks_state: tauri::State<ConversionTasksState>,
     app: tauri::AppHandle,
@@ -312,7 +312,7 @@ fn cancel_conversion(tasks_state: tauri::State<ConversionTasksState>) {
 
 /// Get the transform options for a given sink type
 #[tauri::command]
-fn get_transform(filetype: String) -> Result<TransformerRegistry, Error> {
+fn get_transform(filetype: String) -> Result<TransformerSettings, Error> {
     let sink_provider = select_sink_provider(&filetype).ok_or_else(|| {
         let msg = format!("Invalid sink type: {}", filetype);
         log::error!("{}", msg);
